Make HAS_CPU boolean lazy, speed up import time

[ghstack-poisoned]
